pyproject.toml
[project]
name = "slit"
version = "0.1.0"
description = ""
authors = [{ name = "Jamie Steiner", email = "jvsteiner@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = ["marker-pdf (>=1.6.2,<2.0.0)"]

[tool.poetry]
name = "slit"
version = "0.1.0"
description = ""
authors = ["Jamie Steiner <jvsteiner@gmail.com>"]
package-mode = false                              # Disable package mode

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
streamlit = "*"
marker-pdf = ">=1.6.2,<2.0.0"
asyncio = "*"


[tool.poetry.group.dev.dependencies]
streamlit = "^1.44.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
